LPS-90249 Merge OSGiAllInProcessor into BndDeploymentScenarioGenerator to remove ApplicationArchiveProcessor usages
